Reorganize Azure e2e tests.
