+++ tests/scalers/azure/azure_application_insights/azure_application_insights_test.go
@@ -20,7 +20,7 @@ import (
 )
 
 // Load environment variables from .env file
-var _ = godotenv.Load("../../.env")
+var _ = godotenv.Load("../../../.env")
 
 const (
 	testName = "azure-app-insights-test"
