+++ tests/scalers/azure/azure_blob/azure_blob_test.go
@@ -1,17 +1,17 @@
 //go:build e2e
 // +build e2e
 
-package azure_queue_test
+package azure_blob_test
 
 import (
 	"context"
 	"encoding/base64"
 	"fmt"
 	"os"
+	"strings"
 	"testing"
-	"time"
 
-	"github.com/Azure/azure-storage-queue-go/azqueue"
+	"github.com/Azure/azure-storage-blob-go/azblob"
 	"github.com/joho/godotenv"
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
@@ -24,10 +24,10 @@ import (
 )
 
 // Load environment variables from .env file
-var _ = godotenv.Load("../../.env")
+var _ = godotenv.Load("../../../.env")
 
 const (
-	testName = "azure-queue-test"
+	testName = "azure-blob-test"
 )
 
 var (
@@ -36,7 +36,7 @@ var (
 	secretName       = fmt.Sprintf("%s-secret", testName)
 	deploymentName   = fmt.Sprintf("%s-deployment", testName)
 	scaledObjectName = fmt.Sprintf("%s-so", testName)
-	queueName        = fmt.Sprintf("%s-queue-%d", testName, GetRandomNumber())
+	containerName    = fmt.Sprintf("%s-container-%d", testName, GetRandomNumber())
 )
 
 type templateData struct {
@@ -45,7 +45,7 @@ type templateData struct {
 	Connection       string
 	DeploymentName   string
 	ScaledObjectName string
-	QueueName        string
+	ContainerName    string
 }
 
 const (
@@ -77,14 +77,23 @@ spec:
       labels:
         app: {{.DeploymentName}}
     spec:
+      nodeSelector:
+        kubernetes.io/os: linux
       containers:
         - name: {{.DeploymentName}}
-          image: ghcr.io/kedacore/tests-azure-queue
+          image: slurplk/blob-consumer:latest
           resources:
           env:
             - name: FUNCTIONS_WORKER_RUNTIME
-              value: node
+              value: dotnet
+            - name: AzureFunctionsWebHost__hostid
+              value: {{.DeploymentName}}
             - name: AzureWebJobsStorage
+              valueFrom:
+                secretKeyRef:
+                  name: {{.SecretName}}
+                  key: AzureWebJobsStorage
+            - name: TEST_STORAGE_CONNECTION_STRING
               valueFrom:
                 secretKeyRef:
                   name: {{.SecretName}}
@@ -100,25 +109,26 @@ metadata:
 spec:
   scaleTargetRef:
     name: {{.DeploymentName}}
-  pollingInterval: 5
+  pollingInterval: 10
   minReplicaCount: 0
-  maxReplicaCount: 1
+  maxReplicaCount: 2
   cooldownPeriod: 10
   triggers:
-    - type: azure-queue
+    - type: azure-blob
       metadata:
-        queueName: {{.QueueName}}
+        blobContainerName: {{.ContainerName}}
+        blobCount: '1'
+        activationBlobCount: '5'
         connectionFromEnv: AzureWebJobsStorage
-        activationQueueLength: "5"
 `
 )
 
 func TestScaler(t *testing.T) {
 	// setup
 	t.Log("--- setting up ---")
-	require.NotEmpty(t, connectionString, "AZURE_STORAGE_CONNECTION_STRING env variable is required for azure queue test")
+	require.NotEmpty(t, connectionString, "AZURE_STORAGE_CONNECTION_STRING env variable is required for azure blob test")
 
-	queueURL, messageURL := createQueue(t)
+	containerURL := createContainer(t)
 
 	// Create kubernetes resources
 	kc := GetKubernetesClient(t)
@@ -130,33 +140,31 @@ func TestScaler(t *testing.T) {
 		"replica count should be 0 after 1 minute")
 
 	// test scaling
-	testActivation(t, kc, messageURL)
-	testScaleUp(t, kc, messageURL)
-	testScaleDown(t, kc, messageURL)
+	testActivation(t, kc, containerURL)
+	testScaleUp(t, kc, containerURL)
+	testScaleDown(t, kc, containerURL)
 
 	// cleanup
 	DeleteKubernetesResources(t, kc, testNamespace, data, templates)
-	cleanupQueue(t, queueURL)
+	cleanupContainer(t, containerURL)
 }
 
-func createQueue(t *testing.T) (azqueue.QueueURL, azqueue.MessagesURL) {
-	// Create Queue
+func createContainer(t *testing.T) azblob.ContainerURL {
+	// Create Blob Container
 	httpClient := kedautil.CreateHTTPClient(DefaultHTTPTimeOut, false)
-	credential, endpoint, err := azure.ParseAzureStorageQueueConnection(
+	credential, endpoint, err := azure.ParseAzureStorageBlobConnection(
 		context.Background(), httpClient, kedav1alpha1.AuthPodIdentity{Provider: kedav1alpha1.PodIdentityProviderNone},
 		connectionString, "", "")
 	assert.NoErrorf(t, err, "cannot parse storage connection string - %s", err)
 
-	p := azqueue.NewPipeline(credential, azqueue.PipelineOptions{})
-	serviceURL := azqueue.NewServiceURL(*endpoint, p)
-	queueURL := serviceURL.NewQueueURL(queueName)
+	p := azblob.NewPipeline(credential, azblob.PipelineOptions{})
+	serviceURL := azblob.NewServiceURL(*endpoint, p)
+	containerURL := serviceURL.NewContainerURL(containerName)
 
-	_, err = queueURL.Create(context.Background(), azqueue.Metadata{})
-	assert.NoErrorf(t, err, "cannot create storage queue - %s", err)
+	_, err = containerURL.Create(context.Background(), azblob.Metadata{}, azblob.PublicAccessContainer)
+	assert.NoErrorf(t, err, "cannot create blob container - %s", err)
 
-	messageURL := queueURL.NewMessagesURL()
-	t.Logf("Queue %s created", queueName)
-	return queueURL, messageURL
+	return containerURL
 }
 
 func getTemplateData() (templateData, []Template) {
@@ -168,7 +176,7 @@ func getTemplateData() (templateData, []Template) {
 			Connection:       base64ConnectionString,
 			DeploymentName:   deploymentName,
 			ScaledObjectName: scaledObjectName,
-			QueueName:        queueName,
+			ContainerName:    containerName,
 		}, []Template{
 			{Name: "secretTemplate", Config: secretTemplate},
 			{Name: "deploymentTemplate", Config: deploymentTemplate},
@@ -176,42 +184,53 @@ func getTemplateData() (templateData, []Template) {
 		}
 }
 
-func testActivation(t *testing.T, kc *kubernetes.Clientset, messageURL azqueue.MessagesURL) {
+func testActivation(t *testing.T, kc *kubernetes.Clientset, containerURL azblob.ContainerURL) {
 	t.Log("--- testing activation ---")
-	addMessages(t, messageURL, 3)
-
+	addFiles(t, containerURL, 4)
 	AssertReplicaCountNotChangeDuringTimePeriod(t, kc, deploymentName, testNamespace, 0, 60)
 }
 
-func testScaleUp(t *testing.T, kc *kubernetes.Clientset, messageURL azqueue.MessagesURL) {
+func testScaleUp(t *testing.T, kc *kubernetes.Clientset, containerURL azblob.ContainerURL) {
 	t.Log("--- testing scale up ---")
-	addMessages(t, messageURL, 5)
-
-	assert.True(t, WaitForDeploymentReplicaReadyCount(t, kc, deploymentName, testNamespace, 1, 60, 1),
-		"replica count should be 1 after 1 minute")
+	addFiles(t, containerURL, 10)
+	assert.True(t, WaitForDeploymentReplicaReadyCount(t, kc, deploymentName, testNamespace, 2, 60, 1),
+		"replica count should be 2 after 1 minute")
 }
 
-func testScaleDown(t *testing.T, kc *kubernetes.Clientset, messageURL azqueue.MessagesURL) {
+func testScaleDown(t *testing.T, kc *kubernetes.Clientset, containerURL azblob.ContainerURL) {
 	t.Log("--- testing scale down ---")
-	_, err := messageURL.Clear(context.Background())
-	assert.NoErrorf(t, err, "cannot clear queue - %s", err)
+
+	for i := 0; i < 10; i++ {
+		blobName := fmt.Sprintf("blob-%d", i)
+		blobURL := containerURL.NewBlockBlobURL(blobName)
+
+		_, err := blobURL.Delete(context.Background(), azblob.DeleteSnapshotsOptionInclude,
+			azblob.BlobAccessConditions{})
+
+		assert.NoErrorf(t, err, "cannot delete blob - %s", err)
+	}
 
 	assert.True(t, WaitForDeploymentReplicaReadyCount(t, kc, deploymentName, testNamespace, 0, 60, 1),
 		"replica count should be 0 after 1 minute")
 }
 
-func addMessages(t *testing.T, messageURL azqueue.MessagesURL, count int) {
+func addFiles(t *testing.T, containerURL azblob.ContainerURL, count int) {
+	data := "Hello World!"
+
 	for i := 0; i < count; i++ {
-		msg := fmt.Sprintf("Message - %d", i)
-		_, err := messageURL.Enqueue(context.Background(), msg, 0*time.Second, time.Hour)
-		assert.NoErrorf(t, err, "cannot enqueue message - %s", err)
-		t.Logf("Message queued")
+		blobName := fmt.Sprintf("blob-%d", i)
+		blobURL := containerURL.NewBlockBlobURL(blobName)
+
+		_, err := blobURL.Upload(context.Background(), strings.NewReader(data),
+			azblob.BlobHTTPHeaders{ContentType: "text/plain"}, azblob.Metadata{}, azblob.BlobAccessConditions{},
+			azblob.DefaultAccessTier, nil, azblob.ClientProvidedKeyOptions{}, azblob.ImmutabilityPolicyOptions{})
+
+		assert.NoErrorf(t, err, "cannot upload blob - %s", err)
 	}
 }
 
-func cleanupQueue(t *testing.T, queueURL azqueue.QueueURL) {
+func cleanupContainer(t *testing.T, containerURL azblob.ContainerURL) {
 	t.Log("--- cleaning up ---")
-	_, err := queueURL.Delete(context.Background())
-	assert.NoErrorf(t, err, "cannot delete storage queue - %s", err)
-	t.Logf("Queue %s deleted", queueName)
+	_, err := containerURL.Delete(context.Background(), azblob.ContainerAccessConditions{})
+	assert.NoErrorf(t, err, "cannot delete storage container - %s", err)
 }
