+++ tests/scalers/azure/azure_log_analytics/azure_log_analytics_test.go
@@ -18,7 +18,7 @@ import (
 )
 
 // Load environment variables from .env file
-var _ = godotenv.Load("../../.env")
+var _ = godotenv.Load("../../../.env")
 
 const (
 	testName = "azure-log-analytics-test"
