+++ tests/scalers/azure/azure_pipelines/azure_pipelines_test.go
@@ -24,7 +24,7 @@ import (
 )
 
 // Load environment variables from .env file
-var _ = godotenv.Load("../../.env")
+var _ = godotenv.Load("../../../.env")
 
 const (
 	testName = "azure-pipelines-test"
