+++ tests/scalers/azure/azure_service_bus_topic/azure_service_bus_topic_test.go
@@ -1,7 +1,7 @@
 //go:build e2e
 // +build e2e
 
-package azure_service_bus_queue_test
+package azure_service_bus_topic_test
 
 import (
 	"context"
@@ -21,10 +21,10 @@ import (
 )
 
 // Load environment variables from .env file
-var _ = godotenv.Load("../../.env")
+var _ = godotenv.Load("../../../.env")
 
 const (
-	testName = "azure-service-bus-queue-test"
+	testName = "azure-service-bus-topic-test"
 )
 
 var (
@@ -34,7 +34,8 @@ var (
 	deploymentName   = fmt.Sprintf("%s-deployment", testName)
 	triggerAuthName  = fmt.Sprintf("%s-ta", testName)
 	scaledObjectName = fmt.Sprintf("%s-so", testName)
-	queueName        = fmt.Sprintf("%s-queue-%d", testName, GetRandomNumber())
+	topicName        = fmt.Sprintf("%s-topic-%d", testName, GetRandomNumber())
+	subscriptionName = fmt.Sprintf("%s-subscription-%d", testName, GetRandomNumber())
 )
 
 type templateData struct {
@@ -44,7 +45,8 @@ type templateData struct {
 	DeploymentName   string
 	TriggerAuthName  string
 	ScaledObjectName string
-	QueueName        string
+	TopicName        string
+	SubscriptionName string
 }
 
 const (
@@ -111,7 +113,8 @@ spec:
   triggers:
   - type: azure-servicebus
     metadata:
-      queueName: {{.QueueName}}
+      topicName: {{.TopicName}}
+      subscriptionName: {{.SubscriptionName}}
       activationMessageCount: "5"
     authenticationRef:
       name: {{.TriggerAuthName}}
@@ -123,7 +126,7 @@ func TestScaler(t *testing.T) {
 	t.Log("--- setting up ---")
 	require.NotEmpty(t, connectionString, "AZURE_SERVICE_BUS_CONNECTION_STRING env variable is required for service bus tests")
 
-	client, adminClient := setupServiceBusQueue(t)
+	client, adminClient := setupServiceBusTopicAndSubscription(t)
 
 	kc := GetKubernetesClient(t)
 	data, templates := getTemplateData()
@@ -140,18 +143,20 @@ func TestScaler(t *testing.T) {
 
 	// cleanup
 	DeleteKubernetesResources(t, kc, testNamespace, data, templates)
-	cleanupServiceBusQueue(t, adminClient)
+	cleanupServiceBusTopic(t, adminClient)
 }
 
-func setupServiceBusQueue(t *testing.T) (*azservicebus.Client, *admin.Client) {
+func setupServiceBusTopicAndSubscription(t *testing.T) (*azservicebus.Client, *admin.Client) {
 	adminClient, err := admin.NewClientFromConnectionString(connectionString, nil)
 	assert.NoErrorf(t, err, "cannot connect to service bus namespace - %s", err)
 
-	// Delete the queue if already exists
-	_, _ = adminClient.DeleteQueue(context.Background(), queueName, nil)
+	// Delete the topic if already exists
+	_, _ = adminClient.DeleteTopic(context.Background(), topicName, nil)
 
-	_, err = adminClient.CreateQueue(context.Background(), queueName, nil)
-	assert.NoErrorf(t, err, "cannot create the queue - %s", err)
+	_, err = adminClient.CreateTopic(context.Background(), topicName, nil)
+	assert.NoErrorf(t, err, "cannot create the topic - %s", err)
+	_, err = adminClient.CreateSubscription(context.Background(), topicName, subscriptionName, nil)
+	assert.NoErrorf(t, err, "cannot create the subscription - %s", err)
 
 	client, err := azservicebus.NewClientFromConnectionString(connectionString, nil)
 	assert.NoErrorf(t, err, "cannot connect to service bus namespace - %s", err)
@@ -169,7 +174,8 @@ func getTemplateData() (templateData, []Template) {
 			DeploymentName:   deploymentName,
 			TriggerAuthName:  triggerAuthName,
 			ScaledObjectName: scaledObjectName,
-			QueueName:        queueName,
+			TopicName:        topicName,
+			SubscriptionName: subscriptionName,
 		}, []Template{
 			{Name: "secretTemplate", Config: secretTemplate},
 			{Name: "deploymentTemplate", Config: deploymentTemplate},
@@ -180,7 +186,7 @@ func getTemplateData() (templateData, []Template) {
 
 func testActivation(t *testing.T, kc *kubernetes.Clientset, client *azservicebus.Client) {
 	t.Log("--- testing activation ---")
-	addMessages(t, client, 3)
+	addMessages(t, client, 4)
 
 	AssertReplicaCountNotChangeDuringTimePeriod(t, kc, deploymentName, testNamespace, 0, 60)
 }
@@ -196,17 +202,19 @@ func testScaleUp(t *testing.T, kc *kubernetes.Clientset, client *azservicebus.Cl
 func testScaleDown(t *testing.T, kc *kubernetes.Clientset, adminClient *admin.Client) {
 	t.Log("--- testing scale down ---")
 
-	_, err := adminClient.DeleteQueue(context.Background(), queueName, nil)
-	assert.NoErrorf(t, err, "cannot delete the queue - %s", err)
-	_, err = adminClient.CreateQueue(context.Background(), queueName, nil)
-	assert.NoErrorf(t, err, "cannot create the queue - %s", err)
+	_, err := adminClient.DeleteTopic(context.Background(), topicName, nil)
+	assert.NoErrorf(t, err, "cannot delete the topic - %s", err)
+	_, err = adminClient.CreateTopic(context.Background(), topicName, nil)
+	assert.NoErrorf(t, err, "cannot create the topic - %s", err)
+	_, err = adminClient.CreateSubscription(context.Background(), topicName, subscriptionName, nil)
+	assert.NoErrorf(t, err, "cannot create the subscription - %s", err)
 
 	assert.True(t, WaitForDeploymentReplicaReadyCount(t, kc, deploymentName, testNamespace, 0, 60, 1),
 		"replica count should be 0 after 1 minute")
 }
 
 func addMessages(t *testing.T, client *azservicebus.Client, count int) {
-	sender, err := client.NewSender(queueName, nil)
+	sender, err := client.NewSender(topicName, nil)
 	assert.NoErrorf(t, err, "cannot create the sender - %s", err)
 	for i := 0; i < count; i++ {
 		msg := fmt.Sprintf("Message - %d", i)
@@ -216,8 +224,8 @@ func addMessages(t *testing.T, client *azservicebus.Client, count int) {
 	}
 }
 
-func cleanupServiceBusQueue(t *testing.T, adminClient *admin.Client) {
+func cleanupServiceBusTopic(t *testing.T, adminClient *admin.Client) {
 	t.Log("--- cleaning up ---")
-	_, err := adminClient.DeleteQueue(context.Background(), queueName, nil)
-	assert.NoErrorf(t, err, "cannot delete service bus queue - %s", err)
+	_, err := adminClient.DeleteTopic(context.Background(), topicName, nil)
+	assert.NoErrorf(t, err, "cannot delete service bus topic - %s", err)
 }
